ACF > ACF with Components
I've abstracted the 'Flexi Feilds' out in to elements / components with the help of the 'Component Field Addon' ( http://goo.gl/waIn8X ). Now elements are imported where needed  in to other field groups and can be maintained as seporate field groups, which also helps with max_vars in php.ini files by keeping the fields to a much lower number.